Fix validation bugs and add test files
