+++ lib/validate.js
@@ -23,8 +23,8 @@ var isPositiveInteger = require( 'validate.io-positive-integer' );
 * @returns {Error|Null} error object or null
 */
 function validate( opts, options ) {
-	if ( !isObject( opts ) ) {
-		throw new TypeError( 'invalid input argument. Options argument must be an object. Value: `' + opts + '`.' );
+	if ( !isObject( options ) ) {
+		return new TypeError( 'invalid input argument. Options argument must be an object. Value: `' + options + '`.' );
 	}
 	if ( options.hasOwnProperty( 'level' ) ) {
 		opts.level = options.level;
